+++ contracts/interfaces/IUniswapV2Router.sol
@@ -2,10 +2,12 @@ pragma solidity >=0.6.0 <0.7.0;
 
 interface IUniswapV2Router {
     function swapExactTokensForTokens(
-      uint amountIn,
-      uint amountOutMin,
-      address[] calldata path,
-      address to,
-      uint deadline
-    ) external returns (uint[] memory amounts); 
+        uint256 amountIn,
+        uint256 amountOutMin,
+        address[] calldata path,
+        address to,
+        uint256 deadline
+    )
+        external
+        returns (uint256[] memory amounts);
 }
