+++ contracts/mocks/IUniswapV2RouterMock.sol
@@ -2,13 +2,14 @@ pragma solidity >=0.6.0 <0.7.0;
 
 interface IUniswapV2RouterMock {
     function swapExactTokensForTokens(
-      uint amountIn,
-      uint amountOutMin,
-      address[] calldata path,
-      address to,
-      uint deadline
-    ) external returns (uint[] memory amounts); 
+        uint256 amountIn,
+        uint256 amountOutMin,
+        address[] calldata path,
+        address to,
+        uint256 deadline
+    )
+        external
+        returns (uint256[] memory amounts);
 
-    function getAmountsOut(uint amountIn, address[] memory path) external view returns (uint[] memory amounts);
-
+    function getAmountsOut(uint256 amountIn, address[] memory path) external view returns (uint256[] memory amounts);
 }
